+++ user-service/src/test/java/com/ticketrush/boundedcontext/user/app/usecase/SignupUseCaseTest.java
@@ -12,7 +12,6 @@ import static org.mockito.Mockito.verify;
 import com.ticketrush.boundedcontext.user.app.dto.request.SignupRequest;
 import com.ticketrush.boundedcontext.user.app.dto.response.SignupResponse;
 import com.ticketrush.boundedcontext.user.app.mapper.SignupMapper;
-import com.ticketrush.boundedcontext.user.app.usecase.SignupUseCase;
 import com.ticketrush.boundedcontext.user.domain.entity.User;
 import com.ticketrush.boundedcontext.user.domain.entity.UserAccount;
 import com.ticketrush.boundedcontext.user.out.apiclient.AuthRestClient;
