+++ user-service/src/test/java/com/ticketrush/boundedcontext/user/in/api/v1/UserControllerTest.java
@@ -11,7 +11,7 @@ import com.ticketrush.boundedcontext.user.app.facade.UserFacade;
 import com.ticketrush.global.config.CustomSecurityProperties;
 import com.ticketrush.global.config.InternalApiTokenFilter;
 import com.ticketrush.global.config.SecurityConfig;
-import com.ticketrush.global.security.GatewayHeaderFilter;
+import com.ticketrush.global.filter.GatewayHeaderFilter;
 import com.ticketrush.support.WebMvcSliceTest;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
